timetable-main/api/requirements.txt
pydantic~=1.8.2
fastapi~=0.66.0
requests~=2.25.1
PyYAML~=5.4
uvicorn~=0.14.0
SQLAlchemy~=1.3.14